Added requested changes
Added a button to show a dummy notification.
Changed it so if the user has set a notification, the time displayed matches that.
Added a toast message to confirm to the user that his notification has been sat.
Changed the call to and implementation of CancelNotification to match the desired framework.