updated ability to add category
I tested the "add category" page. It can add a category and add strings under that category. For now everything is in one page as it's not a complete fuly featured page. May need to separate add category and add text into each category, so that it looks less like a test widget.